+++ src/influencefunctional/hybridization/hybridization.jl
@@ -3,3 +3,5 @@ include("partialif/partialif.jl")
 include("ttiif/ttiif.jl")
 
 include("exact_ttiif/exact_ttiif.jl")
+
+include("tdvpif/tdvpif.jl")